ARUHA-681 Refactor TopicRepositoryException
