+++ src/test/java/org/zalando/nakadi/repository/TopicRepositoryHolderTest.java
@@ -7,7 +7,7 @@ import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.domain.Storage;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.exceptions.NakadiRuntimeException;
-import org.zalando.nakadi.exceptions.TopicRepositoryException;
+import org.zalando.nakadi.exceptions.runtime.TopicRepositoryException;
 
 import java.util.List;
